// templates/src/app/interfaces/lines.ts
export interface Lines {
    regular_lines: string[];
    regular_temporary_lines: string[];
    fast_temporary_lines: string[];
    fast_lines: string[];
    zone_lines: string[];
    zone_temporary_lines: string[];
    special_lines: string[];
    cementary_lines: string[];
    express_lines: string[];
    local_lines: string[];
    night_lines: string[];
    substitute_lines: string[];
}